backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "salesshortcut-ai"
version = "0.1.0"
description = "AI-powered Sales Development Representative platform using Portia SDK"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SalesShortcut Team", email = "team@salesshortcut.ai"},
]
keywords = ["sales", "ai", "automation", "sdr", "lead-generation", "portia"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "python-docx",
    "PyPDF2",
    "pandas",
    "pydantic>=2.0.0",
    "portia-sdk-python[all]==0.7.2",
    "requests",
    "beautifulsoup4",
    "flask",
    "uvicorn",
    "fastapi",
    "flask_cors",
    "tavily-python",  # For web search and lead discovery
    "elevenlabs",     # For AI voice calls
    "sendgrid",       # For email outreach
    "sqlalchemy",     # For database management
    "alembic",        # For database migrations
    "python-dotenv",  # For environment variables
    "celery",         # For background tasks
    "redis",          # For task queue
    "schedule",       # For task scheduling
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
    "mypy",
]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/yourusername/github-portia-ai"
Repository = "https://github.com/yourusername/github-portia-ai"
Issues = "https://github.com/yourusername/github-portia-ai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
